refactor-app: Home, Splash, Util 도메인 리팩토링
