--- application/lib/common/exception/failure.dart
@@ -1,8 +1,0 @@
-abstract class Failure {
-  final String message;
-  Failure({required this.message});
-}
-
-abstract class BusinessFailure extends Failure {
-  BusinessFailure({required super.message});
-}
