--- application/lib/common/repository/remote_api.dart
@@ -1,12 +1,0 @@
-
-import 'package:dio/dio.dart';
-import 'package:retrofit/retrofit.dart';
-
-part 'remote_api.g.dart';
-
-@RestApi()
-abstract class RemoteApi {
-
-  factory RemoteApi(Dio dio, {String baseUrl}) = _RemoteApi;
-
-}
