--- application/lib/common/repository/remote_api_provider.dart
@@ -1,16 +1,0 @@
-
-import 'package:riverpod_annotation/riverpod_annotation.dart';
-
-import '../database/dio_provider.dart';
-import 'remote_api.dart';
-
-part 'remote_api_provider.g.dart';
-
-@riverpod
-RemoteApi remoteApi(RemoteApiRef ref) {
-
-  final dio = ref.watch(dioProvider);
-
-  return RemoteApi(dio);
-
-}
