+++ application/lib/features/authentication/domain/entity/oauth_provider.dart
