+++ application/lib/features/authentication/presentation/provider/auth_state_notifier.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:application/features/authentication/domain/entity/auth_entity.dart';
 import 'package:application/features/authentication/domain/provider/auth_provider.dart';
 import 'package:application/features/authentication/domain/use_case/oauth_use_case.dart';
-import 'package:application/features/authentication/presentation/provider/dto/oauth_provider.dart';
+import 'package:application/features/authentication/domain/entity/oauth_provider.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'auth_state_notifier.g.dart';
