+++ application/lib/features/authentication/presentation/screen/auth_screen.dart
@@ -1,5 +1,5 @@
 import 'package:application/features/authentication/presentation/provider/auth_state_notifier.dart';
-import 'package:application/features/authentication/presentation/provider/dto/oauth_provider.dart';
+import 'package:application/features/authentication/domain/entity/oauth_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
