+++ application/lib/features/home/data/model/profile_model.dart
@@ -13,3 +13,4 @@ class ProfileModel with _$ProfileModel {
       _$ProfileModelFromJson(json);
 }
 
+
